+++ src/main/java/com/gromit/gromitmod/gui/button/listener/ScrollListener.java
@@ -4,5 +4,5 @@ import com.gromit.gromitmod.gui.button.AbstractButton;
 
 public interface ScrollListener extends ButtonListener {
 
-    void onListen(AbstractButton abstractButton, int scroll);
+    void onListen(AbstractButton abstractButton, float scroll);
 }
